Prepare 1.0.0 release: remove ontology import, bump version
Drop the OWL/RDF ontology import feature (module, CLI `import` command,
tests, docs, examples, Protégé screenshots) and the now-unused rdflib
dependency. Bump the version to 1.0.0 across Cargo.toml, pyproject.toml,
and the lock files.

+++ tests/compiler_tests/duckdb/23_set_operations.sql
@@ -16,6 +16,6 @@ drop type if exists logicarecord519939597 cascade; create type logicarecord51993
 SELECT
   x_6.unnested_pod AS x
 FROM
-  (select unnest([1, 2, 3, 4, 5]::double[]) as unnested_pod) as x_6, (select unnest([3, 4, 5, 6, 7]::double[]) as unnested_pod) as x_8
+  (select unnest([1, 2, 3, 4, 5]::numeric[]) as unnested_pod) as x_6, (select unnest([3, 4, 5, 6, 7]::numeric[]) as unnested_pod) as x_8
 WHERE
   (x_8.unnested_pod = x_6.unnested_pod) ORDER BY x;
